Josh (#24)
* Create Area.java

This is a compilation of formulas for area, each with its own static function. Feel free to add functions for other shapes.

* Create Volume.java

Compilation of the formulas for volume of solids, each with its own static function. Please feel free to add more functions, especially for prisms.

* Create Motion.java

An abstract class that will serve as a template for motion along a particular axis (development of Unit types is pending).

* Create Unit.java

For easy conversions, though I have yet to implement the conversion functions.

* Update Area.java

Fixed some bugs

* Update Volume.java

Fixed a bug

* Update Area.java

Still fixing bugs

* MassUnit, DistanceUnit, and TimeUnit are now available for use, along with conversions.
Fixed errors in Motion.

* FallingBodyMotion.java is now available on my branch. Still having trouble with Unit.java

* May have resolved issue with Unit.java

* Created LinearMotion.java; renamed some local fields.